Landing improvement, config system
Now a generic script. In-orbit or in-flight activation. Will glide in attempt to reduce error in impact location. Can land at/on a vessel or geocoordinates given as a parameter, or land anywhere. If it cannot reach the destination it will perform a standard landing whereever it is.

Added ship configuration system, enables tweaking of steering manager values of named vessels from text files. May further add properties such as max throttle on engines or RCs, etc.