Various small additions and fixes
The results of the Lambert solver now include the maneuver node needed to complete the transition to the target orbit. The scripts now perform this maneuver in the appropriate situations.

Various small fixes to scripts.

Basic idea implementation of the groundBuilder.

Base implementation for scenarios: the overall string of operations a vehicle will perform in a scenario (e.g. travel to body and land, travel to another and land as a rover, drive to a location).