Created new launch script, added support in missionBuilder
Started a new launch script, partially implemented. Ascent code still needs to be created, but circularization code should work with the Lambert solver.

The missionBuilder will pass appropriate parameters for this script, and should support missions consisting of only a launch (can still pass orbital parameters for it to match).