Launcher and event staging
An improved launcher script. May take different altitudes and a launch profile as parameters, and maintains a TWR of 2 during ascent.

The mission.ks lib declares a local event list variable that events may be added to via addEvent. This list is compared to on every stage_id change, and any appropriate actions are performed.