Event listener, various fixes
The pseudo-event listener was removed and replaced with a proper event listener. Events may be thrown in the regular builder scripts (pre-defined) or in the operation scenario file that is to be executed, using the ON and WHEN commands.

The new event listener now executes code blocks instead of being limited to set functions or right-click functionality. Right-click functionality in event listeners was made easier to interface with in the code blocks.

Events no longer have a TRANSFER_COUNT tagged on, and instead have the name of the craft's parent body.

The kOS_Processor was improved upon and a bug resulting in an additional request file being generated was removed. The source code was also cleaned up.